requirements.txt
Flask
eventlet
flask-socketio
adafruit-circuitpython-motorkit
pyserial